convert: webtoon.js to Typescript
